requirements.txt
sentence-transformers==3.0.1
faiss-cpu==1.8.0
fastapi==0.115.0
uvicorn[standard]==0.30.0
numpy==1.26.4
python-dotenv==1.0.1
scikit-learn==1.5.2
streamlit==1.38.0
